set play control as separate component
