+++ src/components/video/play-control/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
